// src/app/openapi-client/model/userShowDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * DTO zur Anzeige einer Übersicht eines Users
 */
export interface UserShowDto { 
    id?: number;
    firstName?: string;
    lastName?: string;
    email?: string;
    role?: string;
}