// src/app/openapi-client/model/userUpdateDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Aktualisiert Details des Benutzers
 */
export interface UserUpdateDto { 
    firstName?: string;
    lastName?: string;
    email?: string;
    password?: string;
    role?: string;
}