// src/app/openapi-client/model/entityModelUser.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Link } from './link';


export interface EntityModelUser { 
    firstName?: string;
    lastName?: string;
    email?: string;
    password?: string;
    role?: string;
    _links?: { [key: string]: Link; };
}